+++ changelog.md
@@ -129,3 +129,4 @@
 ### 0.?.? - 2021-03-??
 
 * Make `BucketStore` generic for storing workflow files in different types of cloud storage buckets (\#80).
+* Add Google Cloud File Storage bucket implementation (\#80).